+++ .github/workflows/main.yml
@@ -47,8 +47,9 @@ jobs:
         run: poetry install
         if: steps.cache.outputs.cache-hit != 'true'
 
-      - name: Code Quality
-        run: poetry run black . --check
+      # step by step
+      # - name: Code Quality Black
+      #   run: poetry run black . --check
 
       - name: Test with pytest
         run: poetry run pytest --cov . -n 2
